backend/requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Cors==6.0.2
SQLAlchemy==2.0.48
Werkzeug==3.1.6
python-dotenv==1.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
